Move email_helpers to utils/email
