+++ tests/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod ayb;
 pub mod browser;
+pub mod email;
 pub mod testing;
